+++ custom_components/solarcharger/sc_option_state.py
@@ -112,12 +112,6 @@ class ScOptionState(ScConfigState):
     # ----------------------------------------------------------------------------
     # General utils
     # ----------------------------------------------------------------------------
-    def _get_subentry(self, subentry: ConfigSubentry | None) -> ConfigSubentry:
-        if subentry is None:
-            subentry = self._subentry
-        return subentry
-
-    # ----------------------------------------------------------------------------
     def _set_config_value_dict(
         self,
         val_dict: ConfigValueDict | None,
@@ -140,27 +134,21 @@ class ScOptionState(ScConfigState):
         )
 
     # ----------------------------------------------------------------------------
-    def option_get_id(
-        self, config_item: str, subentry: ConfigSubentry | None = None
-    ) -> str | None:
+    def option_get_id(self, config_item: str) -> str | None:
         """Get entity ID from option config data."""
 
-        subentry = self._get_subentry(subentry)
         return get_saved_option_value(
-            self._entry, subentry, config_item, use_default=True
+            self._entry, self._subentry, config_item, use_default=True
         )
 
     # ----------------------------------------------------------------------------
-    def option_get_id_or_abort(
-        self, config_item: str, subentry: ConfigSubentry | None = None
-    ) -> str:
+    def option_get_id_or_abort(self, config_item: str) -> str:
         """Get entity ID from option config data."""
 
-        subentry = self._get_subentry(subentry)
-        entity_id = self.option_get_id(config_item, subentry)
+        entity_id = self.option_get_id(config_item)
         if entity_id is None:
             raise ValueError(
-                f"{subentry.unique_id}: {config_item}: Failed to get entity ID"
+                f"{self._subentry.unique_id}: {config_item}: Failed to get entity ID"
             )
         return entity_id
 
@@ -168,18 +156,16 @@ class ScOptionState(ScConfigState):
     def option_get_string(
         self,
         config_item: str,
-        subentry: ConfigSubentry | None = None,
         val_dict: ConfigValueDict | None = None,
     ) -> str | None:
         """Try to get config from local device settings first, and if not available then try global defaults."""
 
-        subentry = self._get_subentry(subentry)
         str_val = get_saved_option_value(
-            self._entry, subentry, config_item, use_default=True
+            self._entry, self._subentry, config_item, use_default=True
         )
 
         self._set_config_value_dict(
-            val_dict, subentry.unique_id, config_item, None, str_val
+            val_dict, self._subentry.unique_id, config_item, None, str_val
         )
 
         return str_val
@@ -188,12 +174,11 @@ class ScOptionState(ScConfigState):
     def option_get_list(
         self,
         config_item: str,
-        subentry: ConfigSubentry | None = None,
         val_dict: ConfigValueDict | None = None,
     ) -> list[Any] | None:
         """Get list from option config data."""
 
-        json_str = self.option_get_string(config_item, subentry, val_dict)
+        json_str = self.option_get_string(config_item, val_dict=val_dict)
         if json_str is None:
             return None
 
@@ -206,19 +191,17 @@ class ScOptionState(ScConfigState):
     def option_get_entity_number(
         self,
         config_item: str,
-        subentry: ConfigSubentry | None = None,
         val_dict: ConfigValueDict | None = None,
     ) -> float | None:
         """Get entity ID from saved options, then get value for entity."""
         entity_val = None
 
-        subentry = self._get_subentry(subentry)
-        entity_id = self.option_get_id(config_item, subentry)
+        entity_id = self.option_get_id(config_item)
         if entity_id:
             entity_val = self.get_number(entity_id)
 
         self._set_config_value_dict(
-            val_dict, subentry.unique_id, config_item, entity_id, entity_val
+            val_dict, self._subentry.unique_id, config_item, entity_id, entity_val
         )
 
         return entity_val
@@ -227,16 +210,14 @@ class ScOptionState(ScConfigState):
     def option_get_entity_number_or_abort(
         self,
         config_item: str,
-        subentry: ConfigSubentry | None = None,
         val_dict: ConfigValueDict | None = None,
     ) -> float:
         """Get entity ID from saved options, then get value for entity."""
 
-        subentry = self._get_subentry(subentry)
-        entity_val = self.option_get_entity_number(config_item, subentry, val_dict)
+        entity_val = self.option_get_entity_number(config_item, val_dict=val_dict)
         if entity_val is None:
             raise ValueError(
-                f"{subentry.unique_id}: {config_item}: Failed to get entity number value"
+                f"{self._subentry.unique_id}: {config_item}: Failed to get entity number value"
             )
 
         return entity_val
@@ -245,19 +226,17 @@ class ScOptionState(ScConfigState):
     def option_get_entity_integer(
         self,
         config_item: str,
-        subentry: ConfigSubentry | None = None,
         val_dict: ConfigValueDict | None = None,
     ) -> int | None:
         """Get entity name from saved options, then get value for entity."""
         entity_val = None
 
-        subentry = self._get_subentry(subentry)
-        entity_id = self.option_get_id(config_item, subentry)
+        entity_id = self.option_get_id(config_item)
         if entity_id:
             entity_val = self.get_integer(entity_id)
 
         self._set_config_value_dict(
-            val_dict, subentry.unique_id, config_item, entity_id, entity_val
+            val_dict, self._subentry.unique_id, config_item, entity_id, entity_val
         )
 
         return entity_val
@@ -266,19 +245,17 @@ class ScOptionState(ScConfigState):
     def option_get_entity_string(
         self,
         config_item: str,
-        subentry: ConfigSubentry | None = None,
         val_dict: ConfigValueDict | None = None,
     ) -> str | None:
         """Get entity name from saved options, then get value for entity."""
         entity_val = None
 
-        subentry = self._get_subentry(subentry)
-        entity_id = self.option_get_id(config_item, subentry)
+        entity_id = self.option_get_id(config_item)
         if entity_id:
             entity_val = self.get_string(entity_id)
 
         self._set_config_value_dict(
-            val_dict, subentry.unique_id, config_item, entity_id, entity_val
+            val_dict, self._subentry.unique_id, config_item, entity_id, entity_val
         )
 
         return entity_val
@@ -287,19 +264,17 @@ class ScOptionState(ScConfigState):
     def option_get_entity_boolean(
         self,
         config_item: str,
-        subentry: ConfigSubentry | None = None,
         val_dict: ConfigValueDict | None = None,
     ) -> bool | None:
         """Get entity name from saved options, then get value for entity."""
         entity_val = None
 
-        subentry = self._get_subentry(subentry)
-        entity_id = self.option_get_id(config_item, subentry)
+        entity_id = self.option_get_id(config_item)
         if entity_id:
             entity_val = self.get_boolean(entity_id)
 
         self._set_config_value_dict(
-            val_dict, subentry.unique_id, config_item, entity_id, entity_val
+            val_dict, self._subentry.unique_id, config_item, entity_id, entity_val
         )
 
         return entity_val
@@ -308,16 +283,14 @@ class ScOptionState(ScConfigState):
     def option_get_entity_boolean_or_abort(
         self,
         config_item: str,
-        subentry: ConfigSubentry | None = None,
         val_dict: ConfigValueDict | None = None,
     ) -> bool:
         """Get entity ID from saved options, then get value for entity."""
 
-        subentry = self._get_subentry(subentry)
-        entity_val = self.option_get_entity_boolean(config_item, subentry, val_dict)
+        entity_val = self.option_get_entity_boolean(config_item, val_dict)
         if entity_val is None:
             raise ValueError(
-                f"{subentry.unique_id}: {config_item}: Failed to get entity boolean value"
+                f"{self._subentry.unique_id}: {config_item}: Failed to get entity boolean value"
             )
 
         return entity_val
@@ -326,19 +299,17 @@ class ScOptionState(ScConfigState):
     def option_get_entity_time(
         self,
         config_item: str,
-        subentry: ConfigSubentry | None = None,
         val_dict: ConfigValueDict | None = None,
     ) -> time | None:
         """Get entity name from saved options, then get value for entity."""
         entity_val = None
 
-        subentry = self._get_subentry(subentry)
-        entity_id = self.option_get_id(config_item, subentry)
+        entity_id = self.option_get_id(config_item)
         if entity_id:
             entity_val = self.get_time(entity_id)
 
         self._set_config_value_dict(
-            val_dict, subentry.unique_id, config_item, entity_id, entity_val
+            val_dict, self._subentry.unique_id, config_item, entity_id, entity_val
         )
 
         return entity_val
@@ -347,16 +318,14 @@ class ScOptionState(ScConfigState):
     def option_get_entity_time_or_abort(
         self,
         config_item: str,
-        subentry: ConfigSubentry | None = None,
         val_dict: ConfigValueDict | None = None,
     ) -> time:
         """Get entity ID from saved options, then get value for entity."""
 
-        subentry = self._get_subentry(subentry)
-        entity_val = self.option_get_entity_time(config_item, subentry, val_dict)
+        entity_val = self.option_get_entity_time(config_item, val_dict)
         if entity_val is None:
             raise ValueError(
-                f"{subentry.unique_id}: {config_item}: Failed to get entity time value"
+                f"{self._subentry.unique_id}: {config_item}: Failed to get entity time value"
             )
 
         return entity_val
@@ -366,15 +335,13 @@ class ScOptionState(ScConfigState):
         self,
         config_item: str,
         num: float,
-        subentry: ConfigSubentry | None = None,
         val_dict: ConfigValueDict | None = None,
     ) -> None:
         """Set number entity."""
 
-        subentry = self._get_subentry(subentry)
-        entity_id = self.option_get_id(config_item, subentry)
+        entity_id = self.option_get_id(config_item)
         self._set_config_value_dict(
-            val_dict, subentry.unique_id, config_item, entity_id, num
+            val_dict, self._subentry.unique_id, config_item, entity_id, num
         )
 
         if entity_id:
@@ -385,15 +352,13 @@ class ScOptionState(ScConfigState):
         self,
         config_item: str,
         num: int,
-        subentry: ConfigSubentry | None = None,
         val_dict: ConfigValueDict | None = None,
     ) -> None:
         """Set integer entity."""
 
-        subentry = self._get_subentry(subentry)
-        entity_id = self.option_get_id(config_item, subentry)
+        entity_id = self.option_get_id(config_item)
         self._set_config_value_dict(
-            val_dict, subentry.unique_id, config_item, entity_id, num
+            val_dict, self._subentry.unique_id, config_item, entity_id, num
         )
 
         if entity_id:
@@ -403,15 +368,13 @@ class ScOptionState(ScConfigState):
     async def async_option_press_entity_button(
         self,
         config_item: str,
-        subentry: ConfigSubentry | None = None,
         val_dict: ConfigValueDict | None = None,
     ) -> None:
         """Press a button entity."""
 
-        subentry = self._get_subentry(subentry)
-        entity_id = self.option_get_id(config_item, subentry)
+        entity_id = self.option_get_id(config_item)
         self._set_config_value_dict(
-            val_dict, subentry.unique_id, config_item, entity_id
+            val_dict, self._subentry.unique_id, config_item, entity_id
         )
 
         if entity_id:
@@ -422,16 +385,14 @@ class ScOptionState(ScConfigState):
         self,
         config_item: str,
         turn_on: bool,
-        subentry: ConfigSubentry | None = None,
         val_dict: ConfigValueDict | None = None,
     ) -> None:
         """Turn on or off switch entity."""
 
-        subentry = self._get_subentry(subentry)
-        entity_id = self.option_get_id(config_item, subentry)
+        entity_id = self.option_get_id(config_item)
         self._set_config_value_dict(
             val_dict,
-            subentry.unique_id,
+            self._subentry.unique_id,
             config_item,
             entity_id,
             "on" if turn_on else "off",